+++ changelog.md
@@ -6,6 +6,7 @@ Version 18.08
 
 + Add "same link" operator for "VPoint" class.
 + Fix bug of BFGS transform in tinycadlib.
++ Remove test function of Sketch Solve.
 
 Version 18.07
 ===